feature(SimpleSearch): detect if targetNode is an input or a container
Once the targetNode is determined, if is a container, the widget will be append just inside the container.
But if the target is an input element, the widget will clone the input properties and will replace this
input by the composed component.

 - Also tested the clearSearch functionality.
 - Also added "withContainer" to determine if we want an input to be wrapped by a container or not.
 - Also included on SimpleSearch and SuggestedSearch components a "clearSearch" attribute (as a bool)
to determine if we want to appear a clearSearch button when the input value is not empty.